Separate EBML module into files
